+++ truffle/src/com.oracle.truffle.api.debug/src/com/oracle/truffle/api/debug/Breakpoint.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015, 2024, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2015, 2026, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -172,8 +172,6 @@ public class Breakpoint {
          * @since 19.0
          */
         EXCEPTION;
-
-        static final Kind[] VALUES = values();
     }
 
     private static final Breakpoint BUILDER_INSTANCE = new Breakpoint();