+++ truffle/src/com.oracle.truffle.api.object/src/com/oracle/truffle/api/object/DoubleLocation.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013, 2024, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2013, 2026, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -50,13 +50,6 @@ package com.oracle.truffle.api.object;
  */
 @Deprecated(since = "22.2")
 public interface DoubleLocation {
-    /**
-     * @see Location#get(DynamicObject, Shape)
-     * @since 0.8 or earlier
-     */
-    @Deprecated(since = "22.2")
-    double getDouble(DynamicObject store, Shape shape);
-
     /**
      * @see Location#get(DynamicObject, boolean)
      * @since 0.8 or earlier