+++ truffle/src/com.oracle.truffle.api.test/src/com/oracle/truffle/api/test/interop/InteropDefaultsTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, 2024, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2018, 2026, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -42,6 +42,7 @@ package com.oracle.truffle.api.test.interop;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertThrows;
 import static org.junit.Assert.assertTrue;
 
 import java.math.BigDecimal;
@@ -63,6 +64,11 @@ import com.oracle.truffle.api.library.ExportMessage;
 import com.oracle.truffle.api.source.SourceSection;
 
 public class InteropDefaultsTest extends InteropLibraryBaseTest {
+
+    @Test
+    public void testCauseCannotBeInitialized() {
+        assertThrows(IllegalStateException.class, () -> StopIterationException.create().initCause(new RuntimeException()));
+    }
 
     public static class TestInterop1 {
     }